TSDK-802 Add integration test
